requirements.txt
amqp==2.5.2
aniso8601==3.0.2
anyjson==0.3.3
argon2-cffi==19.1.0
asgiref==3.2.3
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.2
decorator==4.3.2
defusedxml==0.5.0
Django==2.2.13
django-allauth==0.41.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.2.0
django-filter==2.1.0
django-graphiql==0.4.4
django-graphql-jwt==0.3.1
django-timezone-field==3.0
django-webpack-loader==0.6.0
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==1.0.2
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.2.1
graphql-relay==2.0.0
html2text==2020.1.16
html5lib==1.0.1
hvac==0.8.2
idna==2.8
importlib-metadata==0.19
jdcal==1.4.1
kombu==4.6.8
mollie-api-python==2.3.1
more-itertools==7.2.0
mysqlclient==1.4.1
oauthlib==3.0.1
openpyxl==3.0.3
Pillow==7.1.2
promise==2.2.1
pycparser==2.19
PyJWT==1.7.1
Pyphen==0.9.5
python-crontab==2.3.8
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
text-unidecode==1.2
tinycss2==1.0.2
urllib3==1.25.3
validators==0.12.6
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
zipp==0.6.0
